New problems Assignment solutions and codeforces stuff

+++ ds-algo/Week_4/majority.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
     int n,found=0;
-    unordered_map<int,int> mp;
+    unordered_map<long int, int> mp;
     cin >> n;
     long long int a[n];
     for(int i=0;i<n;i++)
@@ -13,7 +13,7 @@ int main()
         cin >> a[i];
         mp[a[i]]++;
     }
-    unordered_map<int,int>::iterator it;
+    unordered_map<long int,int>::iterator it;
     for(it = mp.begin();it != mp.end();it++)
     {
         if(it->second > n/2)
